Destroy old home page. It exists in git, after all.
Also, it was causing style problems for the main pages. If I were better
at Grunt and this build:css shit maybe I could have figured it out. Oh
well. It's probably better to relegate all of that to history, anyhow.

I also made the images smaller and made them work with a straight-html
setup.